Add api to docker
