--- cmd/ocp-classroom-api/tracer.go
@@ -1,44 +1,0 @@
-package main
-
-import (
-	"github.com/uber/jaeger-client-go"
-	"github.com/uber/jaeger-lib/metrics"
-
-	jaegercfg "github.com/uber/jaeger-client-go/config"
-	jaegerlog "github.com/uber/jaeger-client-go/log"
-)
-
-// Tip for me: run jaeger in docker and watch trace of MultiCreateClassroomV1 rpc
-// at Jaeger UI http://localhost:16686/search
-
-func InitGlobalTracer(serviceName string) error {
-	// Sample configuration for testing. Use constant sampling to sample every trace
-	// and enable LogSpan to log every span via configured Logger.
-	cfg := jaegercfg.Configuration{
-		Sampler: &jaegercfg.SamplerConfig{
-			Type:  jaeger.SamplerTypeConst,
-			Param: 1,
-		},
-		Reporter: &jaegercfg.ReporterConfig{
-			LogSpans: true,
-		},
-	}
-
-	// Example logger and metrics factory. Use github.com/uber/jaeger-client-go/log
-	// and github.com/uber/jaeger-lib/metrics respectively to bind to real logging and metrics
-	// frameworks.
-	jLogger := jaegerlog.StdLogger
-	jMetricsFactory := metrics.NullFactory
-
-	// Initialize tracer with a logger and a metrics factory
-	_, err := cfg.InitGlobalTracer(
-		serviceName,
-		jaegercfg.Logger(jLogger),
-		jaegercfg.Metrics(jMetricsFactory),
-	)
-	if err != nil {
-		return err
-	}
-
-	return nil
-}
