+++ docker-compose.yml
@@ -1,4 +1,4 @@
-version: "3"
+version: "3.9"
 
 services:
   zookeeper:
@@ -10,7 +10,6 @@ services:
       ZOOKEEPER_CLIENT_PORT: 32181
       ZOOKEEPER_TICK_TIME: 2000
       ZOOKEEPER_SYNC_LIMIT: 2
-
 
   kafka:
     image: confluentinc/cp-kafka
@@ -25,13 +24,10 @@ services:
       KAFKA_LISTENER_SECURITY_PROTOCOL_MAP: INTERNAL:PLAINTEXT,OUTSIDE:PLAINTEXT
       KAFKA_INTER_BROKER_LISTENER_NAME: INTERNAL
 
-      
   jaeger:
     image: jaegertracing/all-in-one:latest
     ports:
-      - "127.0.0.1:6831:6831/udp"
       - "16686:16686"
-
 
   prometheus:
     image: prom/prometheus
@@ -39,3 +35,12 @@ services:
       - "9090:9090"
     volumes:
       - "./prometheus.yml:/etc/prometheus/prometheus.yml"
+
+  ocp-classroom-api:
+    build: .
+    environment: 
+      JAEGER_AGENT_HOST: jaeger
+      JAEGER_AGENT_PORT: 6831
+    ports: 
+      - "5432:5432"
+      - "7002:7002"
