+++ prometheus.yml
@@ -4,4 +4,4 @@ global:
 scrape_configs:
   - job_name: 'ocp-classroom-api'
     static_configs:
-      - targets: ['host.docker.internal:9100']
+      - targets: ['ocp-classroom-api:9100']
